Knives can now be yanked out. SL tracker now updates when SL spawns in. Removed shrapnel scream spam.

See merge request cmdevs/ColonialMarines!3439

+++ code/modules/projectiles/updated_projectiles/gun_helpers.dm
@@ -397,6 +397,8 @@ should be alright.
 	else attachable_overlays[slot] = null
 
 /obj/item/weapon/gun/proc/update_mag_overlay()
+	if(!attachable_overlays["mag"])
+		return
 	var/image/I = attachable_overlays["mag"]
 	overlays -= I
 	qdel(I)
